- predefined colors are extracted to a new directory
- predefined styles are extracted to a new directory

- getStyle function that finds and returns the predefined style (if exists) is added

- the issue with not checking the TEXTGLYPH style before checking for other styles when it comes to extracting default font features is fixed

- the issue with getReactionsLineWidth function if fixed
