+++ src/bindings/python/ctypes/libsbmlnetwork.py.cmake
@@ -5168,6 +5168,17 @@ class LibSBMLNetwork:
         """
         return lib.c_api_setReactionsBorderWidth(self.sbml_object, ctypes.c_double(border_width), layout_index)
 
+    def getReactionsLineWidth(self):
+        """
+        Returns the default line width of the ReactionGlyph objects in the given SBMLDocument
+
+        :Returns:
+
+            a float that determines the default line width of the ReactionGlyph objects in the given SBMLDocument
+        """
+        lib.c_api_getReactionsLineWidth.restype = ctypes.c_double
+        return lib.c_api_getReactionsLineWidth(self.sbml_object)
+
     def setReactionsLineWidth(self, line_width, layout_index=0):
         """
         Sets the line width of all the ReactionGlyph object with the given layout_index in the given SBMLDocument
@@ -9789,16 +9800,28 @@ class LibSBMLNetwork:
 
         return list_of_styles
 
-    def setStyle(self, style_name, layout_index=0):
+    def getStyle(self, renderIndex = 0):
         """
-        Set the a predefined style for the styles of the GraphicalObjects
+        Get the default style of the GraphicalObjects in the layout
+
+        :Returns:
+
+            a string that determines the name of the default style for the styles of the GraphicalObjects
+
+        """
+        lib.c_api_getStyle.restype = ctypes.c_char_p
+        return lib.c_api_getStyle(self.sbml_object, renderIndex).decode()
+
+    def setStyle(self, style_name, render_index=0):
+        """
+        Set the a predefined style of the GraphicalObjects in the layout
 
         :Parameters:
 
             - style_name (string): a string that determines the name of the predefined style for the styles of the GraphicalObjects
 
         """
-        if lib.c_api_setStyle(self.sbml_object, str(style_name).encode(), layout_index) == 0:
+        if lib.c_api_setStyle(self.sbml_object, str(style_name).encode(), render_index) == 0:
             self.enableDisplayReactionsTextLabel(self.whetherDisplayReactionTextLabel(style_name))
             return 0;
